Add Intake enum, switch calculation to use value derived from switch.
The Intake enum represents the two different states that a user's input
can be in. This is a bit more "advanced Swift," but since it can either
be a "direct" milliliter value, or a calorie-based value, those are the
two different conditions. If the user inputs a calorie-based value, we
get the milliliter value by dividing the user's input value by the
number of calories per milliliter, as defined by the Formula subclass
(which you'll see in Formula.milliliters()). We then use this
"milliliters" value in the calculations, which I'll be changing up here
shortly.